Creating tests for nk files to be parsed in.
Creating a test suite for parsing .nk files.
Added a testcase for a single nuke OCIOCDLTransform node with values
normal,
Added a testcase for a single nuke OCIOCDLTransform node with some of
the values not being a float3,
Added a testcase for a single nuke OCIOCDLTransform node with some of
the values being Nuke default (not showing up in the code),
Added a testcase for a multiple nuke OCIOCDLTransform nodes in the same
nk file.